feat: Add animation on homepage and update skills list
This commit removes unused SVG files and adds an animation on the homepage. It also updates the skills list on the Start page by including additional skills such as Express-js, Node-js, and Python. Additionally, it improves the image display in the ProjectCardDesign component, adds an About section to the homepage, and makes various style fixes and updates to the homepage layout and styles. It also addresses warnings for JavaScript disabled, updates the background color of the carousel, and makes changes to card padding and the Container component.